remove nav customisation
Restores the nav in insider as the new theme has changes that makes our customisations
incompatible. But at the same time the theme evolved to make our customisations useless
as it now has the behaviour that we wanted and implemented in the overrides.
The ethical ads part is also useless as Besu pays RTD for no ads.
Makes it simpler now.

Signed-off-by: Nicolas MASSART <nicolas.massart@consensys.net>